fix: generated faulty xcframeworks in assemble.sh
Was previously using the simulator & device specific frameworks that were generated while building `SmartInvesting.xcworkspace` for the respective destinations. This worked on the smart invetsing ipa, but not on others consuming these xcframeworks.
Now we are archiving the respective SDK framework targets and generating xcframeworks from the frameworks present in the archives.

Rename `create_xc_frameworks` to `embed_xc_frameworks`

Refactor the xcframeworks paths of the SDKs

Rename `z-install-scli` to `z-install`. Add `Install xcpretty` step.