fix(ci): add podspec generation for bitrise cocoa deployments
Previously we were maintaining individual podspecs for each remix of the sdk. This was cumbersome and resulted in ambiguity for keeping the versions in sync. We have moved to a system where we have one source of truth from the `Pod.podspec` and generate the corresponding `.podspec.json` using `pod ipc spec`. This allows us to dynamically set the spec values based on our release environment and use the generated spec for deployment.

Remove `update-podspec` GHA

Move the AWS bucket values to inside the `podspec` logic instead of capturing the values from ENV

Stop overriding podspec path on bitrise